refactor into preferences.qr
