+++ collect_app/src/main/AndroidManifest.xml
@@ -127,7 +127,7 @@ the specific language governing permissions and limitations under the License.
             android:configChanges="orientation|screenSize" />
         <activity android:name=".activities.InstanceUploaderActivity" />
         <activity android:name=".activities.AboutActivity" />
-        <activity android:name=".preferences.QRCodeTabsActivity" />
+        <activity android:name=".preferences.qr.QRCodeTabsActivity" />
         <activity android:name=".preferences.PreferencesActivity" />
         <activity android:name=".preferences.AdminPreferencesActivity" />
         <activity android:name=".preferences.FormMetadataPreferencesActivity" />
