+++ collect_app/src/main/java/org/odk/collect/android/preferences/AdminPreferencesFragment.java
@@ -33,6 +33,7 @@ import android.widget.EditText;
 import org.odk.collect.android.R;
 import org.odk.collect.android.fragments.dialogs.MovingBackwardsDialog;
 import org.odk.collect.android.fragments.dialogs.SimpleDialog;
+import org.odk.collect.android.preferences.qr.QRCodeTabsActivity;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
 import org.odk.collect.android.utilities.MultiClickGuard;
