+++ collect_app/src/main/java/org/odk/collect/android/preferences/qr/QRScannerFragment.java
@@ -11,7 +11,7 @@
  * the License.
  */
 
-package org.odk.collect.android.fragments;
+package org.odk.collect.android.preferences.qr;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -29,7 +29,7 @@ import com.journeyapps.barcodescanner.BarcodeResult;
 import com.journeyapps.barcodescanner.DecoratedBarcodeView;
 
 import org.odk.collect.android.R;
-import org.odk.collect.android.preferences.QRCodeTabsActivity;
+import org.odk.collect.android.preferences.utilities.SettingsUtils;
 import org.odk.collect.android.utilities.CompressionUtils;
 import org.odk.collect.android.utilities.ToastUtils;
 
@@ -69,7 +69,7 @@ public class QRScannerFragment extends Fragment implements DecoratedBarcodeView.
                     public void barcodeResult(BarcodeResult result) {
                         beepManager.playBeepSoundAndVibrate();
                         try {
-                            QRCodeTabsActivity.applySettings(getActivity(), CompressionUtils.decompress(result.getText()));
+                            SettingsUtils.applySettings(getActivity(), CompressionUtils.decompress(result.getText()));
                         } catch (IOException | DataFormatException | IllegalArgumentException e) {
                             Timber.e(e);
                             ToastUtils.showShortToast(getString(R.string.invalid_qrcode));
